The application should be reoriented to be multifunctional, a new page with image compression has been added. * Code improvements; * Structure improvements; * Bug fixes.

+++ src/main/java/model/utility/Item.java
@@ -1,5 +1,5 @@
 package model.utility;
 
-public record Item(int id, String title) {
+public record Item(float id, String title) {
 
 }
